requirements.txt
asgiref==3.2.7
Django==3.0.6
django-environ==0.4.5
django-extensions==2.2.9
django-modeltranslation==0.15
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
